Split Python interpreter in two classes
